+++ src/webapp/const.py
@@ -1,0 +1,5 @@
+import os
+
+base = os.path.split(os.path.abspath(__file__))[0]
+call = os.getcwd()
+
